+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/example-demo-app
 
+## 2.8.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.25
+  - @stackframe/stack@2.8.25
+  - @stackframe/stack-ui@2.8.25
+
+## 2.8.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.24
+  - @stackframe/stack-ui@2.8.24
+  - @stackframe/stack@2.8.24
+
 ## 2.8.23
 
 ### Patch Changes