+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/init-stack
 
+## 2.8.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.25
+
+## 2.8.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.24
+
 ## 2.8.23
 
 ### Patch Changes